Add changes suggested by Calton via Slack
https://dcistudents.slack.com/archives/C07342VEX5H/p1716309175300029

I had to make the following changes in order to make it work:
- The method you named mapRolesToAuthorities is an abstract method named
  getAuthorities in the UserDetails interface. I changed the method name
  accordingly.
- The return statement in the renamed getAuthorities method references
  roles which is a property of the a user object. This is a private
  property and can't be accessed directly outside of the class. I added
  a call to user.getRoles() to retrieve the roles.
- The findByUserEmail method returns a Optional<User> object. I added
  get() calls to access the values inside of the optional.

The loadUserByUsername method returns a new User object. The first two
arguments in the constructor, refer to values retrieved from by
findByUserEmail method of the UserRepository interface. The last of
the argument is attempting to call getAuthorities. This is a method of
the SecurityUser class. IntelliJ is not able to resolve that method
call because we don't have a SecurityUser object.