Add Email and ChangePasswordRequest classes
The Email class is used to in the forgot-password-email-form.html
and reset_password_form.html pages to be able to pass an object to
the views.

ChangePasswordRequest is used to accept a current, new and a
confirmation password. It also contains the logic which is used
to check if a new password matches the confirmation password.