Refactor registration endpoints
Commit 2cec6f79 enabled authentication for all requests. This blocked
prevented users from seeing the pages that are part of the registration
flow.

This refactor moved all endpoints that are part of the registration
flow under a common '/register' endpoint. This makes it easier to
match the requests using '/register/**'.

The registration_form.html was renamed register_form.html to match the
the first part of register_success.html so they are grouped together
in the project view.