Add username to offcanvas and header
Used a controller advice to make a string that holds the name
of the currently logged in user available in all views.

The getUser() method, in the ProfileController class, was changed to
have package level access. This method is used in the CommonDataController
to assign the first and last name to a string.

The value of this string is then accessed in the footer and header
using Thymeleaf.