Fix title and age not saving on edit profile page
These two fields were not saving their data to the database when the
form was submitted.

The title had a 'disabled selected' property that was preventing the
default value from being selectable when the page loaded. This caused
the "Mr" title to be selected each time the page loaded and saved,
if the user didn't made another selection.

The age field was not being saved because it was missing the "name"
attribute.