Fix null principal error when user logs out
The exception was caused is by the method in the CommonDataController
which in turn calls the getUser() method in the ProfileController class
to get the current user to get the first and last name.

The getAttributeName method is called whenever any controller method
is called because the class is annotated with @ControllerAdvice.

The problem is that when the user is logged out, the principal is null.
The error that was originally displayed doesn't show that. It only
a 'UserNotFoundException'.

Another way to obtain the name (in our case email address of the
current user) is to pass a Principal object to the
getAttributeName(Principal principal) method. Using this method didn't
solve the problem but an exception was thrown showing the principal
is null after you log out.

The problem was fixed by only allowing the getUserByEmail method to
be called  if the principal is not null.