Bypass field validation on reset_password_form.html
Commit 8662fb7 introduced functionality that allows a user to change
their password. It added validation to the ChangePasswordRequest class.

This class is used by the reset password workflow but it only used
two fields; 'newPassword' and 'confirmationPassword'. The 'password'
field was left unused. When the validation was added to the class
it prevented the reset password form from being submitted because
it failed the validation step.

A dummy password is now set when the reset_password_form.html is
loaded, which allows the validation to pass.

An input of type 'hidden' is also required on the form in order
bypass validation.